refactor: Replace ProfileNavigation into a Component
